♻️ refactor code
