+++ apps/web/app/account/page.tsx
@@ -18,7 +18,7 @@ import { useEffect, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import Head from "../../components/page-components/index/Head";
+import Head from "../../components/page-components/index/Header";
 
 const FormSchema = z.object({
   username: z.string().min(1, {
