+++ apps/web/config/index.ts
@@ -1,3 +1,13 @@
+const path = require("path");
+const dotenv = require("dotenv");
+
+const envName = process.env.NODE_ENV === "production" ? ".env.prod" : ".env";
+
+dotenv.configDotenv({
+  path: path.resolve(__dirname, `../../${envName}`),
+});
+
 export default {
-  serverUrl: "120.0.0.1",
+  isDev: true,
+  serverUrl: process.env.SERVER_URL || "http://120.0.0.1",
 };
